requirements.txt
lerobot[gpu]
ipykernel
ipywidgets
transformers
accelerate
torch
num2words
python-dotenv